+++ src/routes/index.js
@@ -79,7 +79,6 @@ function isStaticRoute(routePattern) {
  * @param {Object} sandboxAuditController - The sandbox audit controller.
  * @param {Object} reportsController - The reports controller.
  * @param {Object} pta2Controller - The PTA2 controller.
- * @param {Object} pta2Controller - The PTA2 controller.
  * @return {{staticRoutes: {}, dynamicRoutes: {}}} - An object with static and dynamic routes.
  */
 export default function getRouteHandlers(
@@ -124,10 +123,14 @@ export default function getRouteHandlers(
     'GET /audits/latest/:auditType': auditsController.getAllLatest,
     'GET /configurations': configurationController.getAll,
     'GET /configurations/latest': configurationController.getLatest,
-    'PUT /configurations/latest': configurationController.updateConfiguration,
+    'PATCH /configurations/latest': configurationController.updateConfiguration,
+    'POST /configurations/:version/restore': configurationController.restoreVersion,
     'GET /configurations/:version': configurationController.getByVersion,
     'POST /configurations/audits': configurationController.registerAudit,
     'DELETE /configurations/audits/:auditType': configurationController.unregisterAudit,
+    'PUT /configurations/latest/queues': configurationController.updateQueues,
+    'PATCH /configurations/latest/jobs/:jobType': configurationController.updateJob,
+    'PATCH /configurations/latest/handlers/:handlerType': configurationController.updateHandler,
     'PATCH /configurations/sites/audits': sitesAuditsToggleController.execute,
     'POST /event/fulfillment': fulfillmentController.processFulfillmentEvents,
     'POST /event/fulfillment/:eventType': fulfillmentController.processFulfillmentEvents,
@@ -336,6 +339,8 @@ export default function getRouteHandlers(
     'GET /sites/:siteId/reports/:reportId': reportsController.getReport,
     'PATCH /sites/:siteId/reports/:reportId': reportsController.patchReport,
     'DELETE /sites/:siteId/reports/:reportId': reportsController.deleteReport,
+
+    'GET /sites-resolve': sitesController.resolveSite,
   };
 
   // Initialization of static and dynamic routes